Bugfix #605 - pull exceptions out of services package
