+++ server/src/main/java/com/objectcomputing/checkins/exceptions/AlreadyExistsException.java
@@ -1,4 +1,4 @@
-package com.objectcomputing.checkins.services.exceptions;
+package com.objectcomputing.checkins.exceptions;
 
 public class AlreadyExistsException extends RuntimeException {
     public AlreadyExistsException(String message) {
