+++ server/src/main/java/com/objectcomputing/checkins/exceptions/BadArgException.java
@@ -1,4 +1,4 @@
-package com.objectcomputing.checkins.services.exceptions;
+package com.objectcomputing.checkins.exceptions;
 
 public class BadArgException extends RuntimeException {
     public BadArgException(String message) {
