+++ server/src/main/java/com/objectcomputing/checkins/exceptions/NotFoundException.java
@@ -1,4 +1,4 @@
-package com.objectcomputing.checkins.services.exceptions;
+package com.objectcomputing.checkins.exceptions;
 
 public class NotFoundException extends RuntimeException {
     public NotFoundException(String message) {
