+++ server/src/main/java/com/objectcomputing/checkins/exceptions/PermissionException.java
@@ -1,4 +1,4 @@
-package com.objectcomputing.checkins.services.exceptions;
+package com.objectcomputing.checkins.exceptions;
 
 public class PermissionException extends RuntimeException {
     public PermissionException(String message) {
