+++ server/src/main/java/com/objectcomputing/checkins/exceptions/handlers/AlreadyExistsHandler.java
@@ -1,9 +1,10 @@
-package com.objectcomputing.checkins.services.exceptions.handlers;
+package com.objectcomputing.checkins.exceptions.handlers;
 
-import com.objectcomputing.checkins.services.exceptions.BadArgException;
+import com.objectcomputing.checkins.exceptions.AlreadyExistsException;
 import io.micronaut.context.annotation.Requires;
 import io.micronaut.http.HttpRequest;
 import io.micronaut.http.HttpResponse;
+import io.micronaut.http.HttpStatus;
 import io.micronaut.http.annotation.Produces;
 import io.micronaut.http.hateoas.JsonError;
 import io.micronaut.http.hateoas.Link;
@@ -13,15 +14,15 @@ import javax.inject.Singleton;
 
 @Produces
 @Singleton
-@Requires(classes = {BadArgException.class, ExceptionHandler.class})
-public class ArgumentsHandler implements ExceptionHandler<BadArgException, HttpResponse> {
+@Requires(classes = {AlreadyExistsException.class, ExceptionHandler.class})
+public class AlreadyExistsHandler implements ExceptionHandler<AlreadyExistsException, HttpResponse> {
 
     @Override
-    public HttpResponse<?> handle(HttpRequest request, BadArgException e) {
+    public HttpResponse<?> handle(HttpRequest request, AlreadyExistsException e) {
         JsonError error = new JsonError(e.getMessage())
                 .link(Link.SELF, Link.of(request.getUri()));
 
-        return HttpResponse.<JsonError>badRequest().body(error);
+        return HttpResponse.<JsonError>status(HttpStatus.CONFLICT).body(error);
     }
 
 }
