+++ server/src/main/java/com/objectcomputing/checkins/exceptions/handlers/ArgumentsHandler.java
@@ -1,0 +1,27 @@
+package com.objectcomputing.checkins.exceptions.handlers;
+
+import com.objectcomputing.checkins.exceptions.BadArgException;
+import io.micronaut.context.annotation.Requires;
+import io.micronaut.http.HttpRequest;
+import io.micronaut.http.HttpResponse;
+import io.micronaut.http.annotation.Produces;
+import io.micronaut.http.hateoas.JsonError;
+import io.micronaut.http.hateoas.Link;
+import io.micronaut.http.server.exceptions.ExceptionHandler;
+
+import javax.inject.Singleton;
+
+@Produces
+@Singleton
+@Requires(classes = {BadArgException.class, ExceptionHandler.class})
+public class ArgumentsHandler implements ExceptionHandler<BadArgException, HttpResponse> {
+
+    @Override
+    public HttpResponse<?> handle(HttpRequest request, BadArgException e) {
+        JsonError error = new JsonError(e.getMessage())
+                .link(Link.SELF, Link.of(request.getUri()));
+
+        return HttpResponse.<JsonError>badRequest().body(error);
+    }
+
+}
