+++ server/src/main/java/com/objectcomputing/checkins/exceptions/handlers/PermissionsHandler.java
@@ -1,9 +1,10 @@
-package com.objectcomputing.checkins.services.exceptions.handlers;
+package com.objectcomputing.checkins.exceptions.handlers;
 
-import com.objectcomputing.checkins.services.exceptions.NotFoundException;
+import com.objectcomputing.checkins.exceptions.PermissionException;
 import io.micronaut.context.annotation.Requires;
 import io.micronaut.http.HttpRequest;
 import io.micronaut.http.HttpResponse;
+import io.micronaut.http.HttpStatus;
 import io.micronaut.http.annotation.Produces;
 import io.micronaut.http.hateoas.JsonError;
 import io.micronaut.http.hateoas.Link;
@@ -13,15 +14,15 @@ import javax.inject.Singleton;
 
 @Produces
 @Singleton
-@Requires(classes = {NotFoundException.class, ExceptionHandler.class})
-public class NotFoundHandler implements ExceptionHandler<NotFoundException, HttpResponse> {
+@Requires(classes = {PermissionException.class, ExceptionHandler.class})
+public class PermissionsHandler implements ExceptionHandler<PermissionException, HttpResponse> {
 
     @Override
-    public HttpResponse<?> handle(HttpRequest request, NotFoundException e) {
+    public HttpResponse<?> handle(HttpRequest request, PermissionException e) {
         JsonError error = new JsonError(e.getMessage())
                 .link(Link.SELF, Link.of(request.getUri()));
 
-        return HttpResponse.<JsonError>notFound().body(error);
+        return HttpResponse.status(HttpStatus.FORBIDDEN).body(error);
     }
 
 }
