+++ server/src/main/java/com/objectcomputing/checkins/services/guild/member/GuildMemberServicesImpl.java
@@ -1,6 +1,6 @@
 package com.objectcomputing.checkins.services.guild.member;
 
-import com.objectcomputing.checkins.services.exceptions.BadArgException;
+import com.objectcomputing.checkins.exceptions.BadArgException;
 import com.objectcomputing.checkins.services.guild.GuildRepository;
 import com.objectcomputing.checkins.services.memberprofile.MemberProfileRepository;
