+++ server/src/main/java/com/objectcomputing/checkins/services/memberprofile/MemberProfileServicesImpl.java
@@ -3,7 +3,7 @@ package com.objectcomputing.checkins.services.memberprofile;
 
 import com.objectcomputing.checkins.security.InsufficientPrivelegesException;
 import com.objectcomputing.checkins.services.checkins.CheckInServices;
-import com.objectcomputing.checkins.services.exceptions.NotFoundException;
+import com.objectcomputing.checkins.exceptions.NotFoundException;
 import com.objectcomputing.checkins.services.member_skill.MemberSkillServices;
 import com.objectcomputing.checkins.services.memberprofile.currentuser.CurrentUserServices;
 import com.objectcomputing.checkins.services.role.Role;
