+++ server/src/main/java/com/objectcomputing/checkins/services/validate/ArgumentsValidation.java
@@ -1,6 +1,6 @@
 package com.objectcomputing.checkins.services.validate;
 
-import com.objectcomputing.checkins.services.exceptions.BadArgException;
+import com.objectcomputing.checkins.exceptions.BadArgException;
 
 import javax.inject.Singleton;
 import javax.validation.constraints.NotNull;
