+++ server/src/test/java/com/objectcomputing/checkins/services/checkindocument/CheckinDocumentControllerTest.java
@@ -380,7 +380,7 @@ public class CheckinDocumentControllerTest extends TestContainersSuite implement
         HttpClientResponseException responseException = assertThrows(HttpClientResponseException.class, () ->
                 client.toBlocking().exchange(request, Map.class));
 
-        assertEquals(HttpStatus.UNAUTHORIZED, responseException.getStatus());
+        assertEquals(HttpStatus.FORBIDDEN, responseException.getStatus());
         assertEquals("You do not have permission to access this resource", responseException.getMessage());
     }
