+++ server/src/test/java/com/objectcomputing/checkins/services/pulseresponse/PulseResponseServiceImplTest.java
@@ -1,6 +1,6 @@
 package com.objectcomputing.checkins.services.pulseresponse;
 
-import com.objectcomputing.checkins.services.exceptions.BadArgException;
+import com.objectcomputing.checkins.exceptions.BadArgException;
 import com.objectcomputing.checkins.services.memberprofile.MemberProfile;
 import com.objectcomputing.checkins.services.memberprofile.MemberProfileRepository;
 import io.micronaut.test.annotation.MicronautTest;
