+++ server/src/test/java/com/objectcomputing/checkins/services/team/member/TeamMemberControllerTest.java
@@ -489,8 +489,7 @@ class TeamMemberControllerTest extends TestContainersSuite implements TeamFixtur
                 () -> client.toBlocking().exchange(request, Map.class));
 
         assertNotNull(responseException.getResponse());
-        assertEquals(HttpStatus.UNAUTHORIZED, responseException.getStatus());
-
+        assertEquals(HttpStatus.FORBIDDEN, responseException.getStatus());
     }
 
     @Test
